+++ backend/src/main/java/com/example/smu_club/auth/controller/PublicAuthController.java
@@ -6,6 +6,7 @@ import com.example.smu_club.auth.dto.LoginRequest;
 import com.example.smu_club.auth.dto.ReissueRequest;
 import com.example.smu_club.auth.dto.SignupRequest;
 import com.example.smu_club.auth.service.AuthService;
+import com.example.smu_club.auth.service.TokenService;
 import com.example.smu_club.common.ApiResponseDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -24,6 +25,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class PublicAuthController {
 
     private final AuthService authService;
+    private final TokenService tokenService;
 
     // 로그인시도
     @PostMapping("/login")
@@ -39,7 +41,7 @@ public class PublicAuthController {
     // 사용자의 accessToken 이 만료되었을 때 이 API 호출을 통해서 RefreshToken 을 통해서 accessToken 을 재발급
     @PostMapping("/reissue")
     public ResponseEntity<ApiResponseDto<JwtTokenResponse>> reissue(@RequestBody ReissueRequest reissueRequest) {
-        JwtTokenResponse jwtTokenResponse = authService.reissueTokens(reissueRequest);
+        JwtTokenResponse jwtTokenResponse = tokenService.reissueTokens(reissueRequest);
 
         ApiResponseDto<JwtTokenResponse> response = ApiResponseDto.success(jwtTokenResponse, "토큰이 성공적으로 발급되었습니다.");
         return ResponseEntity.ok(response);
